backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
python-multipart==0.0.9
faster-whisper==1.0.3
huggingface-hub==0.23.4
requests==2.32.3
python-dotenv==1.0.1
pydantic==2.7.4